requirements.txt
Django==2.0.3
djangorestframework==3.7.7
djoser==1.1.5
pytz==2018.3
requests==2.18.4
awesome-slugify==1.6.5